Working version with reduced amount on bindings
