+++ src/ccvPreAsm.js
@@ -1,0 +1,2 @@
+// Helpers used in the C++ bindings
+var Module = module.exports = Module || {};
